Add a basic menu to the shell panel
The menu currently displays a "Hello World" message that is badly
centered on multi-display setups.

The icon needs to be changed. I don't know how hard it will be to have
an icon with a dynamic number on it showing the number of unread
messages, but that will be the best solution if it's easy enough.